Split out actors inbox
