+++ activitypub.php
@@ -42,6 +42,7 @@ function rest_init() {
 	Rest\Server::init();
 	Rest\Post::init();
 	( new Rest\Actors_Controller() )->register_routes();
+	( new Rest\Actors_Inbox_Controller() )->register_routes();
 	( new Rest\Application_Controller() )->register_routes();
 	( new Rest\Collections_Controller() )->register_routes();
 	( new Rest\Comments_Controller() )->register_routes();
