+++ tests/includes/rest/class-test-actors-inbox-controller.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Test file for Activitypub Rest Inbox.
+ * Test file for Actors_Inbox_Controller.
  *
  * @package Activitypub
  */
@@ -8,11 +8,12 @@
 namespace Activitypub\Tests\Rest;
 
 /**
- * Test class for Activitypub Rest Inbox.
+ * Test class for Actors_Inbox_Controller.
  *
- * @coversDefaultClass \Activitypub\Rest\Inbox
+ * @group rest
+ * @coversDefaultClass \Activitypub\Rest\Actors_Inbox_Controller
  */
-class Test_Inbox extends \WP_UnitTestCase {
+class Test_Actors_Inbox_Controller extends \Activitypub\Tests\Test_REST_Controller_Testcase {
 	/**
 	 * Test user ID.
 	 *
@@ -29,17 +30,10 @@ class Test_Inbox extends \WP_UnitTestCase {
 
 	/**
 	 * Create fake data before tests run.
-	 *
-	 * @param WP_UnitTest_Factory $factory Helper that creates fake data.
 	 */
-	public static function wpSetUpBeforeClass( $factory ) {
-		self::$user_id = $factory->user->create(
-			array(
-				'role' => 'author',
-			)
-		);
-
-		self::$post_id = $factory->post->create(
+	public static function set_up_before_class() {
+		self::$user_id = self::factory()->user->create( array( 'role' => 'author' ) );
+		self::$post_id = self::factory()->post->create(
 			array(
 				'post_author'  => self::$user_id,
 				'post_title'   => 'Test Post',
@@ -52,7 +46,7 @@ class Test_Inbox extends \WP_UnitTestCase {
 	/**
 	 * Clean up after tests.
 	 */
-	public static function wpTearDownAfterClass() {
+	public static function tear_down_after_class() {
 		wp_delete_user( self::$user_id );
 	}
 
@@ -85,7 +79,7 @@ class Test_Inbox extends \WP_UnitTestCase {
 			'object' => 'https://local.example/@test',
 		);
 
-		$request = new \WP_REST_Request( 'POST', '/activitypub/1.0/users/1/inbox' );
+		$request = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/users/1/inbox' );
 		$request->set_header( 'Content-Type', 'application/activity+json' );
 		$request->set_body( \wp_json_encode( $json ) );
 
@@ -107,7 +101,7 @@ class Test_Inbox extends \WP_UnitTestCase {
 			'actor' => 'https://remote.example/@test',
 		);
 
-		$request = new \WP_REST_Request( 'POST', '/activitypub/1.0/users/1/inbox' );
+		$request = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/users/1/inbox' );
 		$request->set_header( 'Content-Type', 'application/activity+json' );
 		$request->set_body( \wp_json_encode( $json ) );
 
@@ -133,31 +127,7 @@ class Test_Inbox extends \WP_UnitTestCase {
 			'object' => 'https://local.example/@test',
 		);
 
-		$request = new \WP_REST_Request( 'POST', '/activitypub/1.0/users/1/inbox' );
-		$request->set_header( 'Content-Type', 'application/activity+json' );
-		$request->set_body( \wp_json_encode( $json ) );
-
-		// Dispatch the request.
-		$response = \rest_do_request( $request );
-		$this->assertEquals( 202, $response->get_status() );
-
-		\remove_filter( 'activitypub_defer_signature_verification', '__return_true' );
-	}
-
-	/**
-	 * Test follow request global inbox.
-	 */
-	public function test_follow_request_global_inbox() {
-		\add_filter( 'activitypub_defer_signature_verification', '__return_true' );
-
-		$json = array(
-			'id'     => 'https://remote.example/@id',
-			'type'   => 'Follow',
-			'actor'  => 'https://remote.example/@test',
-			'object' => 'https://local.example/@test',
-		);
-
-		$request = new \WP_REST_Request( 'POST', '/activitypub/1.0/inbox' );
+		$request = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/users/1/inbox' );
 		$request->set_header( 'Content-Type', 'application/activity+json' );
 		$request->set_body( \wp_json_encode( $json ) );
 
@@ -182,7 +152,7 @@ class Test_Inbox extends \WP_UnitTestCase {
 			'object' => 'https://local.example/@test',
 		);
 
-		$request = new \WP_REST_Request( 'POST', '/activitypub/1.0/users/1/inbox' );
+		$request = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/users/1/inbox' );
 		$request->set_header( 'Content-Type', 'application/activity+json' );
 		$request->set_body( \wp_json_encode( $json ) );
 
@@ -199,49 +169,7 @@ class Test_Inbox extends \WP_UnitTestCase {
 			'inReplyTo' => 'https://local.example/post/test',
 			'published' => '2020-01-01T00:00:00Z',
 		);
-		$request        = new \WP_REST_Request( 'POST', '/activitypub/1.0/users/1/inbox' );
-		$request->set_header( 'Content-Type', 'application/activity+json' );
-		$request->set_body( \wp_json_encode( $json ) );
-
-		// Dispatch the request.
-		$response = \rest_do_request( $request );
-		$this->assertEquals( 202, $response->get_status() );
-
-		\remove_filter( 'activitypub_defer_signature_verification', '__return_true' );
-	}
-
-	/**
-	 * Test create request global inbox.
-	 */
-	public function test_create_request_global_inbox() {
-		\add_filter( 'activitypub_defer_signature_verification', '__return_true' );
-
-		// Invalid request, because of an invalid object.
-		$json = array(
-			'id'     => 'https://remote.example/@id',
-			'type'   => 'Create',
-			'actor'  => 'https://remote.example/@test',
-			'object' => 'https://local.example/@test',
-		);
-
-		$request = new \WP_REST_Request( 'POST', '/activitypub/1.0/inbox' );
-		$request->set_header( 'Content-Type', 'application/activity+json' );
-		$request->set_body( \wp_json_encode( $json ) );
-
-		// Dispatch the request.
-		$response = \rest_do_request( $request );
-		$this->assertEquals( 400, $response->get_status() );
-		$this->assertEquals( 'rest_invalid_param', $response->get_data()['code'] );
-
-		// Valid request, because of a valid object.
-		$json['object'] = array(
-			'id'        => 'https://remote.example/post/test',
-			'type'      => 'Note',
-			'content'   => 'Hello, World!',
-			'inReplyTo' => 'https://local.example/post/test',
-			'published' => '2020-01-01T00:00:00Z',
-		);
-		$request        = new \WP_REST_Request( 'POST', '/activitypub/1.0/inbox' );
+		$request        = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/users/1/inbox' );
 		$request->set_header( 'Content-Type', 'application/activity+json' );
 		$request->set_body( \wp_json_encode( $json ) );
 
@@ -271,7 +199,7 @@ class Test_Inbox extends \WP_UnitTestCase {
 			),
 		);
 
-		$request = new \WP_REST_Request( 'POST', '/activitypub/1.0/users/1/inbox' );
+		$request = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/users/1/inbox' );
 		$request->set_header( 'Content-Type', 'application/activity+json' );
 		$request->set_body( \wp_json_encode( $json ) );
 
@@ -295,7 +223,7 @@ class Test_Inbox extends \WP_UnitTestCase {
 			'object' => 'https://local.example/post/test',
 		);
 
-		$request = new \WP_REST_Request( 'POST', '/activitypub/1.0/users/1/inbox' );
+		$request = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/users/1/inbox' );
 		$request->set_header( 'Content-Type', 'application/activity+json' );
 		$request->set_body( \wp_json_encode( $json ) );
 
@@ -319,7 +247,7 @@ class Test_Inbox extends \WP_UnitTestCase {
 			'object' => 'https://local.example/post/test',
 		);
 
-		$request = new \WP_REST_Request( 'POST', '/activitypub/1.0/users/1/inbox' );
+		$request = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/users/1/inbox' );
 		$request->set_header( 'Content-Type', 'application/activity+json' );
 		$request->set_body( \wp_json_encode( $json ) );
 
@@ -328,88 +256,6 @@ class Test_Inbox extends \WP_UnitTestCase {
 		$this->assertEquals( 202, $response->get_status() );
 
 		\remove_filter( 'activitypub_defer_signature_verification', '__return_true' );
-	}
-
-	/**
-	 * Test whether an activity is public.
-	 *
-	 * @dataProvider the_data_provider
-	 *
-	 * @param array $data  The data.
-	 * @param bool  $check The check.
-	 */
-	public function test_is_activity_public( $data, $check ) {
-		$this->assertEquals( $check, \Activitypub\is_activity_public( $data ) );
-	}
-
-	/**
-	 * Data provider.
-	 *
-	 * @return array[]
-	 */
-	public function the_data_provider() {
-		return array(
-			array(
-				array(
-					'cc'     => array(
-						'https://example.org/@test',
-						'https://example.com/@test2',
-					),
-					'to'     => 'https://www.w3.org/ns/activitystreams#Public',
-					'object' => array(),
-				),
-				true,
-			),
-			array(
-				array(
-					'cc'     => array(
-						'https://example.org/@test',
-						'https://example.com/@test2',
-					),
-					'to'     => array(
-						'https://www.w3.org/ns/activitystreams#Public',
-					),
-					'object' => array(),
-				),
-				true,
-			),
-			array(
-				array(
-					'cc'     => array(
-						'https://example.org/@test',
-						'https://example.com/@test2',
-					),
-					'object' => array(),
-				),
-				false,
-			),
-			array(
-				array(
-					'cc'     => array(
-						'https://example.org/@test',
-						'https://example.com/@test2',
-					),
-					'object' => array(
-						'to' => 'https://www.w3.org/ns/activitystreams#Public',
-					),
-				),
-				true,
-			),
-			array(
-				array(
-					'cc'     => array(
-						'https://example.org/@test',
-						'https://example.com/@test2',
-					),
-					'object' => array(
-						'to' => array(
-							'https://www.w3.org/ns/activitystreams#Public',
-						),
-					),
-				),
-				true,
-			),
-		);
 	}
 
 	/**
@@ -486,5 +332,54 @@ class Test_Inbox extends \WP_UnitTestCase {
 		$this->assertEquals( 202, $response->get_status() );
 
 		remove_filter( 'pre_get_remote_metadata_by_actor', '__return_true' );
+	}
+
+	/**
+	 * Test schema.
+	 *
+	 * @covers ::get_item_schema
+	 */
+	public function test_get_item_schema() {
+		$request  = new \WP_REST_Request( 'OPTIONS', '/' . ACTIVITYPUB_REST_NAMESPACE . '/actors/1/inbox' );
+		$response = rest_get_server()->dispatch( $request );
+		$schema   = $response->get_data()['schema'];
+
+		$this->assertIsArray( $schema );
+		$this->assertEquals( 'object', $schema['type'] );
+		$this->assertArrayHasKey( 'properties', $schema );
+	}
+
+	/**
+	 * Test that the Followers response matches its schema.
+	 *
+	 * @covers ::get_items
+	 * @covers ::get_item_schema
+	 */
+	public function test_response_matches_schema() {
+		add_filter(
+			'activitypub_rest_inbox_array',
+			function ( $inbox ) {
+				$inbox['totalItems'] = 1;
+
+				return $inbox;
+			}
+		);
+
+		$request  = new \WP_REST_Request( 'GET', '/' . ACTIVITYPUB_REST_NAMESPACE . '/users/1/inbox' );
+		$response = rest_get_server()->dispatch( $request );
+		$data     = $response->get_data();
+		$schema   = ( new \Activitypub\Rest\Actors_Inbox_Controller() )->get_item_schema();
+
+		$valid = \rest_validate_value_from_schema( $data, $schema );
+		$this->assertNotWPError( $valid, 'Response failed schema validation: ' . ( \is_wp_error( $valid ) ? $valid->get_error_message() : '' ) );
+	}
+
+	/**
+	 * Test get_item method.
+	 *
+	 * @doesNotPerformAssertions
+	 */
+	public function test_get_item() {
+		// Controller does not implement get_item().
 	}
 }
