+++ tests/includes/rest/class-test-inbox-controller.php
@@ -1,0 +1,186 @@
+<?php
+/**
+ * Test file for Activitypub Rest Inbox.
+ *
+ * @package Activitypub
+ */
+
+namespace Activitypub\Tests\Rest;
+
+/**
+ * Test class for Activitypub Rest Inbox.
+ *
+ * @group rest
+ * @coversDefaultClass \Activitypub\Rest\Inbox_Controller
+ */
+class Test_Inbox_Controller extends \Activitypub\Tests\Test_REST_Controller_Testcase {
+	/**
+	 * Test follow request global inbox.
+	 *
+	 * @covers ::get_items
+	 */
+	public function test_get_items() {
+		\add_filter( 'activitypub_defer_signature_verification', '__return_true' );
+
+		$json = array(
+			'id'     => 'https://remote.example/@id',
+			'type'   => 'Follow',
+			'actor'  => 'https://remote.example/@test',
+			'object' => 'https://local.example/@test',
+		);
+
+		$request = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/inbox' );
+		$request->set_header( 'Content-Type', 'application/activity+json' );
+		$request->set_body( \wp_json_encode( $json ) );
+
+		// Dispatch the request.
+		$response = \rest_do_request( $request );
+		$this->assertEquals( 202, $response->get_status() );
+
+		\remove_filter( 'activitypub_defer_signature_verification', '__return_true' );
+	}
+
+	/**
+	 * Test create request global inbox.
+	 *
+	 * @covers ::create_item
+	 */
+	public function test_create_item() {
+		\add_filter( 'activitypub_defer_signature_verification', '__return_true' );
+
+		// Invalid request, because of an invalid object.
+		$json = array(
+			'id'     => 'https://remote.example/@id',
+			'type'   => 'Create',
+			'actor'  => 'https://remote.example/@test',
+			'object' => 'https://local.example/@test',
+		);
+
+		$request = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/inbox' );
+		$request->set_header( 'Content-Type', 'application/activity+json' );
+		$request->set_body( \wp_json_encode( $json ) );
+
+		// Dispatch the request.
+		$response = \rest_do_request( $request );
+		$this->assertEquals( 400, $response->get_status() );
+		$this->assertEquals( 'rest_invalid_param', $response->get_data()['code'] );
+
+		// Valid request, because of a valid object.
+		$json['object'] = array(
+			'id'        => 'https://remote.example/post/test',
+			'type'      => 'Note',
+			'content'   => 'Hello, World!',
+			'inReplyTo' => 'https://local.example/post/test',
+			'published' => '2020-01-01T00:00:00Z',
+		);
+		$request        = new \WP_REST_Request( 'POST', '/' . ACTIVITYPUB_REST_NAMESPACE . '/inbox' );
+		$request->set_header( 'Content-Type', 'application/activity+json' );
+		$request->set_body( \wp_json_encode( $json ) );
+
+		// Dispatch the request.
+		$response = \rest_do_request( $request );
+		$this->assertEquals( 202, $response->get_status() );
+
+		\remove_filter( 'activitypub_defer_signature_verification', '__return_true' );
+	}
+
+	/**
+	 * Test whether an activity is public.
+	 *
+	 * @dataProvider public_activity_provider
+	 *
+	 * @param array $data  The data.
+	 * @param bool  $check The check.
+	 */
+	public function test_is_activity_public( $data, $check ) {
+		$this->assertEquals( $check, \Activitypub\is_activity_public( $data ) );
+	}
+
+	/**
+	 * Data provider.
+	 *
+	 * @return array[]
+	 */
+	public function public_activity_provider() {
+		return array(
+			array(
+				array(
+					'cc'     => array(
+						'https://example.org/@test',
+						'https://example.com/@test2',
+					),
+					'to'     => 'https://www.w3.org/ns/activitystreams#Public',
+					'object' => array(),
+				),
+				true,
+			),
+			array(
+				array(
+					'cc'     => array(
+						'https://example.org/@test',
+						'https://example.com/@test2',
+					),
+					'to'     => array(
+						'https://www.w3.org/ns/activitystreams#Public',
+					),
+					'object' => array(),
+				),
+				true,
+			),
+			array(
+				array(
+					'cc'     => array(
+						'https://example.org/@test',
+						'https://example.com/@test2',
+					),
+					'object' => array(),
+				),
+				false,
+			),
+			array(
+				array(
+					'cc'     => array(
+						'https://example.org/@test',
+						'https://example.com/@test2',
+					),
+					'object' => array(
+						'to' => 'https://www.w3.org/ns/activitystreams#Public',
+					),
+				),
+				true,
+			),
+			array(
+				array(
+					'cc'     => array(
+						'https://example.org/@test',
+						'https://example.com/@test2',
+					),
+					'object' => array(
+						'to' => array(
+							'https://www.w3.org/ns/activitystreams#Public',
+						),
+					),
+				),
+				true,
+			),
+		);
+	}
+
+	/**
+	 * Test get_item method.
+	 *
+	 * @doesNotPerformAssertions
+	 */
+	public function test_get_item() {
+		// Controller does not implement get_item().
+	}
+
+	/**
+	 * Test get_item_schema method.
+	 *
+	 * @doesNotPerformAssertions
+	 */
+	public function test_get_item_schema() {
+		// Controller does not implement get_item_schema().
+	}
+}
